+++ clib/disk2file_util.c
@@ -202,7 +202,7 @@ m5_2_disk2file(ptr_in,lclc,lclm,ip) /* return values:
 
      struct disk2file_cmd *lclc;  /* result structure with parameters */
      struct disk2file_mon *lclm;  /* result structure with parameters */
-     long ip[5];   /* standard parameter array */
+     int ip[5];   /* standard parameter array */
 {
   char *new_str, *ptr, *ptr2, *ptr_save;
   int count, ierr;
@@ -319,7 +319,7 @@ m5_scan_set_2_disk2file(ptr_in,lclc,lclm,ip) /* return values:
 
      struct disk2file_cmd *lclc;  /* result structure with parameters */
      struct disk2file_mon *lclm;  /* result structure with parameters */
-     long ip[5];   /* standard parameter array */
+     int ip[5];   /* standard parameter array */
 {
   char *new_str, *ptr, *ptr2, *ptr_save;
   int count, ierr, mk5b;
